Feature10

+++ journalApplication/src/main/java/com/darshan/journalApplication/service/JournalEntryService.java
@@ -5,6 +5,7 @@ import com.darshan.journalApplication.repository.JournalEntryRepository;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -19,6 +20,7 @@ public class JournalEntryService {
     @Autowired
     private UserEntryService userEntryService;
 
+    @Transactional
     public void EntryRecord(JournalEntry journalEntry, String userName) {
             User user = userEntryService.findByUserName(userName);
             journalEntry.setDate(LocalDateTime.now());
@@ -41,11 +43,18 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
+    @Transactional
     public void deleteById(ObjectId id, String userName) {
-        User user = userEntryService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userEntryService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
-
+        try {
+            User user = userEntryService.findByUserName(userName);
+            boolean b = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if(b) {
+                userEntryService.saveEntry(user);
+                journalEntryRepository.deleteById(id);
+        }
+        } catch (Exception e) {
+            System.out.println(e);
+            throw new RuntimeException("An Error Occured while deleting.");
+        }
     }
 }
